test/boost: move scrub tests to a separate file
Move scrub tests, along with helpers dedicated for them, contained in
`sstable_compaction_test.cc` to a dedicated `scrub_test.cc` file. This
separates scrub tests from unrelated tests in
`sstable_compaction_test.cc` and provides a place for automatic scrub
tests to be added in the future.

The tests using `compact_corrupted_by_compression_mode` and scrub were
not moved, as they form a set of test cases with tests using other
compaction
types, which utilize one static helper.
